+++ core/applib/src/main/java/org/apache/isis/applib/AppManifestAbstract2.java
@@ -22,7 +22,7 @@ import java.util.Collections;
 import java.util.List;
 
 import org.apache.isis.applib.annotation.Programmatic;
-import org.apache.isis.testing.fixtures.applib.fixturescripts.FixtureScript;
+import org.apache.isis.applib.fixturescripts.FixtureScript;
 
 /**
  * Convenience adapter, configured using a {@link AppManifestAbstract.Builder}.
